pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["Glaucco Lima <glauccolima@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
factory-boy = "^3.3.1"
pytest = "^8.3.3"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
gitpython = "^3.1.43"
whitenoise = "^6.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
